bug(modal-close)
fix modal even after form submission
