+++ client/components/protected/GroupList.jsx
@@ -2,7 +2,9 @@ import React from 'react';
 import { Modal } from 'react-bootstrap';
 import AppStore from '../../stores/AppStore';
 import Group from './Group';
-import CreateGroup from './CreateGroup';
+import AppActions from '../../actions/AppActions';
+import * as Api from '../../api/AppApi';
+
 
 /**
  * A collection of group that displays the auser's groups
@@ -25,21 +27,64 @@ class GroupList extends React.Component {
     super(props);
     this.state = {
       isOpen: false,
+      groupName: '',
+      error: ''
     };
+    this.onChange = this.onChange.bind(this);
+    this.handleSubmit = this.handleSubmit.bind(this);
     this.closeModal = this.closeModal.bind(this);
     this.openModal = this.openModal.bind(this);
   }
 
+
 /**
-* Handles close Modal event
+* @method onChange
+*
+* @description Monitors changes in the components and change the state
+*
+* @memberof CreateGroup
 *
 * @param {SyntheticEvent} event
 *
-* @returns {void} null
+* @returns {void}
 */
-  closeModal() {
-    this.setState({ isOpen: false });
+  onChange(event) {
+    // event.preventDefault();
+    this.setState({
+      groupName: event.target.value
+    });
   }
+
+
+/**
+ * @description creates a group if groupname is not empty
+ *
+ * @param {object} event
+ *
+ * @returns {void}
+ *
+ * @memberof CreateGroup
+*/
+  handleSubmit(event) {
+    event.preventDefault();
+    const group = {
+      groupName: this.state.groupName,
+    };
+    if ((!this.state.groupName) || (!this.state.groupName.trim())) {
+      this.setState({
+        error: 'Please enter a valid group name'
+      });
+    } else {
+      AppActions.createGroup(group);
+      Api.getUserGroups();
+      this.setState({
+        groupName: '',
+        error: '',
+        isOpen: false
+      });
+    }
+  }
+
 
 /**
 * Handles open Modal event
@@ -48,8 +93,22 @@ class GroupList extends React.Component {
 *
 * @returns {void} null
 */
-  openModal() {
+  openModal(event) {
+    event.preventDefault();
     this.setState({ isOpen: true });
+  }
+
+
+/**
+* Handles close Modal event
+*
+* @param {SyntheticEvent} event
+*
+* @returns {void} null
+*/
+  closeModal(event) {
+    event.preventDefault();
+    this.setState({ isOpen: false });
   }
 
 /**
@@ -83,7 +142,30 @@ class GroupList extends React.Component {
           </button>
           <Modal show={this.state.isOpen} onHide={this.closeModal}>
             <Modal.Body>
-              <CreateGroup />
+              <div className="panel-body">
+                <div className="error"> {this.state.error} </div>
+                <form
+                  onSubmit={this.handleSubmit}
+                  className="navbar-form"
+                  role="search"
+                >
+                  <div className="form-group">
+                    <input
+                      type="text"
+                      className="form-control"
+                      placeholder="Create Group"
+                      name="groupName"
+                      onChange={this.onChange}
+                      value={this.state.groupName}
+                    />
+                  </div>
+                  <button
+                    type="submit"
+                    id="submit"
+                    className="btn btn-primary addMember"
+                  > Submit </button>
+                </form>
+              </div>
             </Modal.Body>
             <Modal.Footer>
               <a onClick={this.closeModal}> Close</a>
