+++ client/components/protected/MessageList.jsx
@@ -1,9 +1,10 @@
 import React from 'react';
+import lodash from 'lodash';
 import { Modal } from 'react-bootstrap';
 import MessageBox from './MessageBox';
-import AddMember from './AddMember';
 import Message from './Message';
 import AppStore from '../../stores/AppStore';
+import AppActions from '../../actions/AppActions';
 
 /**
 * A collection of message that displays the available messages
@@ -26,21 +27,95 @@ class MessageList extends React.Component {
     super(props);
     this.state = {
       isOpen: false,
+      userName: '',
+      userId: '',
+      error: '',
     };
+    this.onChange = this.onChange.bind(this);
+    this.handleSubmit = this.handleSubmit.bind(this);
+    this.changeToUserId = this.changeToUserId.bind(this);
     this.closeModal = this.closeModal.bind(this);
     this.openModal = this.openModal.bind(this);
   }
 
+
 /**
-* Handles close Modal event
+ * @description function that get userid from username
+ *
+ * @param {String} userName
+ *
+ * @returns {String}
+ *
+ * @memberof AddMember
+ */
+  changeToUserId(userName) {
+    let userId;
+    lodash.map(AppStore.getUsers()).map((user) => {
+      if (userName === user.userName) {
+        userId = user.userId;
+      } else {
+        return null;
+      }
+    });
+    return userId;
+  }
+
+
+/**
+* @method onChange
+*
+* @description Monitors changes in the components and change the state
+*
+* @memberof AddMember
 *
 * @param {SyntheticEvent} event
 *
-* @returns {void} void
+* @returns {void}
 */
-  closeModal() {
-    this.setState({ isOpen: false });
+  onChange(event) {
+    this.setState({
+      userName: event.target.value,
+    });
   }
+
+
+/**
+ * @description add a member to a group if a group is selected and
+ * username is not empty
+ *
+ * @param {SyntheticEvent} event
+ *
+ * @returns {void}
+ *
+ * @memberof AddMember
+*/
+  handleSubmit(event) {
+    event.preventDefault();
+    if (!AppStore.getOpenedGroup()[0]) {
+      this.setState({
+        error: 'Kindly select a group first',
+      });
+      return true;
+    }
+    const user = {
+      userId: this.changeToUserId(this.state.userName),
+      userName: this.state.userName,
+      groupId: AppStore.getOpenedGroup()[0].groupId,
+    };
+    if (!user.userId) {
+      this.setState({
+        error: 'This User does not exist',
+      });
+    } else {
+      AppActions.addUserToGroup(user);
+      this.setState({
+        error: '',
+        userName: '',
+        isOpen: false
+      });
+    }
+  }
+
 
 /**
 * Handles Open Modal event
@@ -52,6 +127,19 @@ class MessageList extends React.Component {
   openModal(event) {
     event.preventDefault();
     this.setState({ isOpen: true });
+  }
+
+
+/**
+* Handles close Modal event
+*
+* @param {SyntheticEvent} event
+*
+* @returns {void} void
+*/
+  closeModal(event) {
+    event.preventDefault();
+    this.setState({ isOpen: false });
   }
 
 /**
@@ -96,7 +184,28 @@ class MessageList extends React.Component {
         </div>
         <Modal show={this.state.isOpen} onHide={this.closeModal}>
           <Modal.Body>
-            <AddMember />
+            <div className="panel-body">
+              <h6> To add a member, type the member username </h6>
+              <div className="error"> {this.state.error} </div>
+              <form onSubmit={this.handleSubmit} className="navbar-form" role="search">
+                <div className="form-group">
+                  <input
+                    name="userName"
+                    onChange={this.onChange}
+                    value={this.state.userName}
+                    type="text"
+                    className="form-control"
+                    placeholder="Add member"
+                  />
+                </div>
+                <button
+                  type="submit"
+                  className="btn btn-primary addMember"
+                >
+            Submit </button>
+              </form>
+              <br />
+            </div>
           </Modal.Body>
           <Modal.Footer>
             <a onClick={this.closeModal}> Close</a>
